backend/requirements.txt
fastapi>=0.116.1,<0.117.0
uvicorn>=0.35.0,<0.36.0
httpx>=0.28.1,<0.29.0
python-dotenv>=1.1.1,<2.0.0
pydantic-settings>=2.0.0,<3.0.0
bs4>=0.0.2,<0.0.3
firebase-admin<7.0.0
google-cloud-aiplatform>=1.71.1,<2.0.0
crewai>=0.150.0,<0.151.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
vertexai>=1.71.1,<2.0.0
crawl4ai>=0.7.0,<1.0.0
pytest>=7.0.0,<8.0.0
nest-asyncio>=1.5.0,<2.0.0
rich>=13.7.0,<14.0.0
structlog>=24.1.0,<25.0.0
opentelemetry-api>=1.24.0,<2.0.0
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.45b0,<1.0.0
opentelemetry-instrumentation-requests>=0.45b0,<1.0.0
opentelemetry-instrumentation-logging>=0.45b0,<1.0.0
opentelemetry-exporter-gcp-trace>=1.6.0,<2.0.0
prometheus-client>=0.20.0,<1.0.0
sentry-sdk[fastapi]>=1.40.0,<2.0.0
psutil>=6.1.1
